refactor: organize app into the common template directory structure (#74)
