+++ listeners/actions/index.js
@@ -1,0 +1,5 @@
+import { feedbackActionCallback } from './feedback.js';
+
+export const register = (app) => {
+  app.action('feedback', feedbackActionCallback);
+};
