+++ listeners/events/index.js
@@ -1,0 +1,5 @@
+import { appMentionCallback } from './app_mention.js';
+
+export const register = (app) => {
+  app.event('app_mention', appMentionCallback);
+};
